requirements.txt
asgiref==3.5.2
Django==4.1
django-model-utils==4.2.0
django-quill-editor==0.1.40
django-widget-tweaks==1.4.12
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.2
